requirements.txt
django==2.0.5
djangorestframework==3.8.2
jinja2==2.9.6

